Add support for Dancer2's upcoming change_session_id method.
Session engine must have a `_change_id` method in order for `change_id`
to be called. We need to override `change_id` since
`Plack::Middleware::Session` handles ID itself and new IDs are not
generated by D2.

Also switch to in-memory session handler instead of cookie since we
want to test that ID is being changed.